Updated all knaspsack visualizations and KA exercises by Jeremy Rodencal of UWOshkosh.
Also added Jeremy's N-choose-K slideshow

Finally added Jeremy's initial attempt at an RST file for the Dynamic Programming tutorial material.   Much remain to be done on this, including figuring out how to embed the visualizations into the page.
Problems in doing this may be linked to our better understanding the role of the corresponding xml file.