Gender neurtal pronouns
He/his/him/she/her/hers were all made gender neutral by change to
they/their/them, by rewording to plural for generic people, or to it if
it was not to a person. A handful of gendered pronouns were left if in
reference to a specific person and gender assignment seemed reasonable.
This is the convention for this textbook.