insert & remove examples changed
The examples for skip lists were changed for several reasons.
First, they were made a little shorter. Second, the remove not
reverses the inserts so you see the same lists in reverse.
Third, the adjustHead is not done on the first insert.
Fourth, more cases are shown for nodes before/after being
larger or smaller.