working new C++ List version
This is the start of doing an updated C++ (non-template)version
of the code. Code still needs work.
It is mostly synced with the Java version so they are much more similar.
The test code is very basic and the start of doing way Java does.
Note void * was not used for the types because there are issues
with doing toString in a resonable way.
ErrorRec.h is a code to help with testing and will be used in all
the new test codes.