resolve TODOs in book code (not test code)
Returning ListItemType& creates issues because items were accepted
as const. It also is tricky to return an empty item (currently NULL
which is not great) as a const. Thus, decided not to do this.
Good practice is virtual destructor in base class so leaving.
There are still TODOs for items to add to text.
Minor fix to includes in test code (not needed for now)