remove toString
This commit only does this one item so this could be reverted
if desired. The decisions was made to not support this
function because it is difficult to do correctly in all
languages. In C++ you do not have an easy way to create
a string from any templated type.
Note leaving special my_to_string so can test int
and string since this
is only used in test code and should not be seen
by book readers.