add exceptions to C++ code
The decision was made to use exceptions instead of returning
the null value. In C++ templates it is not easy to return
a non-existant
value since primitives can be used.
Note the same change will soon be made to the Java versions.
It also removes my_to_string from the C++_templates AList version since
it was only used during testing.
Adds test for getValue on empty list.
Adds tests for remove and getValue at end of list.
Modifies test codes to check for exceptions.