Refactored
Previous design broke if the the <div id=Header> existed, but was the first div in the <body> of the page. The call to header_tag.next_sibling.insert_before() would break because insert_before would return None.
New design finds all of the <script> tags in the header div, and then appends them to the <head> tag. then the  header div is removed.