Updated CW exercise names to make them prettier
There is a bug in OpenDSA where in some cases, the gym name defined in CW and the section header where that gym is published in the book need to match. If they don’t match, the Book->LTI widget breaks, and the book isn’t completely build. Updated the names here to make them easier to read by the student, and also match the other CW gyms in the course.