Deactivated in-page shortcuts
The base url for all LTI-hosted pages in canvas is opendsa.cs.vt.edu/lti-launch . When these pages are embedded in Canvas as external links, they are invoked with an HTTP POST command, which the OpenDSA LTI service has a route to handle. If you want to use internal links like the shortcuts in an LTI-hosted page, the URL that is invoked is the same as the page (with the extra hashtag for the internal target) but it is invoked by an HTTP GET request, which the OpenDSA LTI tool doesn’t handle for the lti-launch route. These changes can be reverted to restore the internal shortcuts.