Rename codeinclude accessibility libs
Renamed the CSS and JS files that are supposed to improve screen reader support for the codeinclude blocks. These are not included by default, so there should be any side effects, as no config files have these included. Didn’t delete because the issue still needs sorted, and I’ve not had time to look back at it.